requirements.txt
fastapi
uvicorn
openai-whisper
transformers
torch
mysql-connector-python
pyspellchecker
requests
python-multipart
pdfkit